feat: start update with useq
