+++ .pre-commit-config.yaml
@@ -28,4 +28,4 @@ repos:
         files: "^src/"
         additional_dependencies:
           - pymmcore-plus >=0.10.2
-          - useq-schema >=0.4.7
+          - useq-schema @ git+https://github.com/pymmcore-plus/useq-schema  # to revert once released
