+++ examples/hcs_temp/fov_selector.py
@@ -7,7 +7,8 @@ except ImportError:
 
 from qtpy.QtWidgets import QApplication
 
-from pymmcore_widgets.hcs._fov_widget import Center, FOVSelectorWidget
+from pymmcore_widgets.hcs._fov_widget._fov_sub_widgets import Center
+from pymmcore_widgets.hcs._fov_widget._fov_widget import FOVSelectorWidget
 from pymmcore_widgets.hcs._plate_model import load_database
 
 database_path = (
@@ -18,13 +19,13 @@ database = load_database(database_path)
 
 app = QApplication([])
 
+plate = database["96-well"]
 fs = FOVSelectorWidget(
-    plate=database["standard 96 wp"],
-    mode=Center(x=0, y=0, fov_width=512, fov_height=512),
+    plate=plate, mode=Center(x=0, y=0, fov_width=512, fov_height=512)
 )
 
 fs.valueChanged.connect(lambda x: rich_print(x))
 
 fs.show()
 
-app.exec_()
+app.exec()
