+++ src/pymmcore_widgets/hcs/__init__.py
@@ -1,5 +1,5 @@
 """HCS Wizard."""
 
-from ._main_wizard_widget import HCSWizard
+from ._main_wizard._main_wizard import HCSWizard
 
 __all__ = ["HCSWizard"]
